feat(useInputRefs): ref 커스텀 훅 만들어서 autofocus 기능 수정
